Versucht, in XMLaltTest (alt:alternativ) das initialisieren effizienter zu machen. Keinerlei Garantie, dass es richtig funktioniert, aber es ist schnell... (wird noch nicht in Graphen eingefügt)

+++ src/application/graphNavigation/Graph.java
@@ -16,7 +16,7 @@ public class Graph {
         this.numberOfAddedMotorwayLinks = 0;
     }
 
-    int getMatrixNodeNumberById(int id){
+    int getMatrixNodeNumberById(long id){
         int pruefer = -1;
         for(int i = 0; i < numberOfAddedNodes; i++){
             if(nodes[i] == null){
