+++ src/application/graphNavigation/Node.java
@@ -2,20 +2,20 @@ package application.graphNavigation;
 
 public class Node {
 
-    int id;
+    long id;
 
     boolean link;
     double longitude;
     double latitude;
 
-    public Node (int id, boolean isLink, double lon, double lat) {
+    public Node (long id, boolean isLink, double lon, double lat) {
         this.id = id;
         this.link = isLink;
         this.longitude = lon;
         this.latitude = lat;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
